tests, resolve nested resolution
